Remove map function
I looked through the python documentation some more,
and I realized I was being dumb. You don't need to map
the lambda to every single value. The filter function
can just use the lambda as a rule to evaluate
the range as an iterable and select those for which
it evaluates true. I only realized this potential after
working with the select function in ruby.

I still think that my ruby version has higher potential,
but at least python isn't as bad as I thought.