Replace Loop with Select function
This is already so much better than my python version.
I can see why I always lose to ruby.
Instead of looping through every value and finding the
variables that fit all in one step,
this solution generates the range of numbers using
'..' and selects the elements that evaluate to
true under the provided rule.
It's basically like the filter() method from
python.